new hook: shasumCheck
Unlike other hooks, this one replaces a chunk of functionality. This is
to allow for use-cases where a straight-up shasum check from the
blobstore/filesystem isn't what we want to do. This can also be used to
bypass the shasum check entirely, for non-mirroring use-cases.
