+++ lib/hooks.js
@@ -3,7 +3,7 @@ Copyright (c) 2014, Yahoo! Inc. All rights reserved.
 Code licensed under the BSD License.
 See LICENSE file.
 */
-var options = require('./args.js');
+var options = require('./args');
 var log = require('davlog');
 var hooks = options.hooks;
 
@@ -18,8 +18,12 @@ function defaultHook(data, callback, success) {
 }
 
 function wrap(hookName) {
+    return wrapHook(hooks[hookName]);
+}
+
+function wrapHook(hook) {
     return function wrapped(data, callback, success){
-        hooks[hookName].call(hookContext, data, function(err, good) {
+        hook.call(hookContext, data, function(err, good) {
             if (err) {
                 return callback(err);
             }
@@ -39,3 +43,4 @@ exports.versionJson = hooks.hasOwnProperty('versionJson') ? wrap('versionJson') 
 exports.indexJson = hooks.hasOwnProperty('indexJson') ? wrap('indexJson') : defaultHook;
 exports.globalIndexJson = hooks.hasOwnProperty('globalIndexJson') ? wrap('globalIndexJson') : defaultHook;
 exports.startup = hooks.hasOwnProperty('startup') ? wrap('startup') : defaultHook;
+exports.shasumCheck = hooks.hasOwnProperty('shasumCheck') ? wrap('shasumCheck') : wrapHook(require('./defaultShasumCheck'));
