+++ lib/verify.js
@@ -5,6 +5,7 @@ See LICENSE file.
 */
 var http = require('http-https'),
     options = require('./args'),
+    hooks = require('./hooks'),
     crypto = require('crypto'),
     timethat = require('timethat').calc,
     url = require('url'),
@@ -96,35 +97,15 @@ var verify = function(info, callback) {
             }
             log.info('[' + counter[info.path] + '] checking shasum of', info.tarball);
             process.nextTick(function() {
-                var shasum = crypto.createHash('sha1');
-                shasum.setEncoding('hex');
-                options.blobstore.createReadStream(info.tarball)
-                    .on('error', function(e) {
-                        // There are some totally crap situations where the tarball
-                        // path is something silly, like a directory, but still a
-                        // path on the filesystem, so `exists` passes. Ignore and continue.
-                        log.err(' [' + counter[info.path] + '] failed to read from path which supposedly exists:', info.path);
-                        log.err(e.stack);
-                        callback(null, info);
-                    })
-                    .on('end', function() {
-                        shasum.end();
-                        var d = shasum.read();
-                        if (info.shasum !== d) {
-                            info.error = {
-                                found: d,
-                                expected: info.shasum
-                            };
-                            log.err(' [' + counter[info.path] + '] shasum check failed for', info.path);
-                            log.err(' [' + counter[info.path] + '] found', d, 'expected', info.shasum);
-                            return exports.update(info, callback);
-                        }
-                        log.info('[' + counter[info.path] + '] shasum check passed for', info.path, '(' + info.shasum + ')');
-                        delete counter[info.path];
-                        callback(null, info);
-                    })
-                    .pipe(shasum);
+                hooks.shasumCheck(info, function(){
+                    // shasum failed
+                    exports.update(info, callback);
+                }, function(){
+                    // shasum passed
+                    delete counter[info.path];
+                    callback(null, info);
                 });
+            });
         });
     });
 };
